chore: move struct parsing into submodule
